requirements.txt
requests
numpy
google-api-python-client
google-auth-oauthlib
python-dotenv
-f git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube_dl
youtube_dl
spotipy
flask